@@ -106,60 +106,60 @@
     "zone.js": "0.14.4"
   },
   "devDependencies": {
-    "@angular-builders/custom-webpack": "17.0.1",
-    "@angular-builders/jest": "17.0.2",
-    "@angular-devkit/build-angular": "17.3.0",
+    "@angular-builders/custom-webpack": "17.0.2",
+    "@angular-builders/jest": "17.0.3",
+    "@angular-devkit/build-angular": "17.3.3",
     "@angular-eslint/eslint-plugin": "17.3.0",
-    "@angular/cli": "17.3.0",
-    "@angular/compiler-cli": "17.3.0",
-    "@angular/service-worker": "17.3.0",
+    "@angular/cli": "17.3.3",
+    "@angular/compiler-cli": "17.3.3",
+    "@angular/service-worker": "17.3.3",
     "@types/jest": "29.5.12",
     "@types/node": "20.11.25",
-    "@typescript-eslint/eslint-plugin": "7.3.1",
-    "@typescript-eslint/parser": "7.3.1",
+    "@typescript-eslint/eslint-plugin": "7.6.0",
+    "@typescript-eslint/parser": "7.6.0",
     "browser-sync": "3.0.2",
     "browser-sync-webpack-plugin": "2.3.0",
     "buffer": "6.0.3",
     "concurrently": "8.2.2",
     "copy-webpack-plugin": "12.0.2",
-    "cypress": "13.7.0",
+    "cypress": "13.7.2",
     "cypress-audit": "1.1.0",
     "eslint": "8.57.0",
     "eslint-config-prettier": "9.1.0",
     "eslint-plugin-cypress": "2.15.1",
     "eslint-webpack-plugin": "4.1.0",
     "folder-hash": "4.0.4",
-    "generator-jhipster": "8.2.1",
+    "generator-jhipster": "8.3.0",
     "husky": "9.0.11",
     "jest": "29.7.0",
-    "jest-date-mock": "1.0.8",
+    "jest-date-mock": "1.0.9",
     "jest-environment-jsdom": "29.7.0",
     "jest-junit": "16.0.0",
     "jest-preset-angular": "14.0.3",
     "jest-sonar": "0.2.16",
-    "lighthouse": "11.6.0",
+    "lighthouse": "11.7.0",
     "lint-staged": "15.2.2",
     "merge-jsons-webpack-plugin": "2.0.1",
     "prettier": "3.2.5",
     "prettier-plugin-java": "2.6.0",
-    "prettier-plugin-packagejson": "2.4.12",
+    "prettier-plugin-packagejson": "2.4.14",
     "rimraf": "5.0.5",
-    "swagger-ui-dist": "5.12.0",
+    "swagger-ui-dist": "5.13.0",
     "ts-jest": "29.1.2",
-    "typescript": "5.4.2",
+    "typescript": "5.4.4",
     "wait-on": "7.2.0",
     "webpack-bundle-analyzer": "4.10.1",
     "webpack-merge": "5.10.0",
     "webpack-notifier": "1.15.0"
   },
   "engines": {
-    "node": ">=20.11.1"
+    "node": ">=20.12.1"
   },
   "cacheDirectories": [
     "node_modules"
   ],
   "overrides": {
     "browser-sync": "3.0.2",
-    "webpack": "5.90.3"
+    "webpack": "5.91.0"
   }
 }
