@@ -1,4 +1,4 @@
-import { ChangeDetectionStrategy, Component, Input } from '@angular/core';
+import { ChangeDetectionStrategy, Component, input } from '@angular/core';
 
 import SharedModule from 'app/shared/shared.module';
 import { CacheMetrics } from 'app/admin/metrics/metrics.model';
@@ -15,12 +15,12 @@ export class MetricsCacheComponent {
   /**
    * object containing all cache related metrics
    */
-  @Input() cacheMetrics?: { [key: string]: CacheMetrics };
+  cacheMetrics = input<{ [key: string]: CacheMetrics }>();
 
   /**
    * boolean field saying if the metrics are in the process of being updated
    */
-  @Input() updating?: boolean;
+  updating = input<boolean>();
 
-  filterNaN = (input: number): number => filterNaN(input);
+  filterNaN = (n: number): number => filterNaN(n);
 }
