@@ -6,8 +6,8 @@ import SharedModule from 'app/shared/shared.module';
 import { ActivateService } from './activate.service';
 
 @Component({
-  selector: 'jhi-activate',
   standalone: true,
+  selector: 'jhi-activate',
   imports: [SharedModule, RouterModule],
   templateUrl: './activate.component.html',
 })
