@@ -3,8 +3,8 @@ import { Component, ElementRef, inject, Input, Renderer2 } from '@angular/core';
 import SharedModule from 'app/shared/shared.module';
 
 @Component({
-  selector: 'jhi-password-strength-bar',
   standalone: true,
+  selector: 'jhi-password-strength-bar',
   imports: [SharedModule],
   templateUrl: './password-strength-bar.component.html',
   styleUrl: './password-strength-bar.component.scss',
