@@ -10,8 +10,8 @@ import PasswordStrengthBarComponent from '../password/password-strength-bar/pass
 import { RegisterService } from './register.service';
 
 @Component({
-  selector: 'jhi-register',
   standalone: true,
+  selector: 'jhi-register',
   imports: [SharedModule, RouterModule, FormsModule, ReactiveFormsModule, PasswordStrengthBarComponent],
   templateUrl: './register.component.html',
 })
