@@ -2,7 +2,6 @@ import { ComponentFixture, TestBed, fakeAsync, inject, tick } from '@angular/cor
 import { HttpHeaders, HttpResponse } from '@angular/common/http';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { ActivatedRoute } from '@angular/router';
-import { RouterTestingModule } from '@angular/router/testing';
 import { of, Subject } from 'rxjs';
 import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 
@@ -20,11 +19,7 @@ describe('Authority Management Component', () => {
 
   beforeEach(() => {
     TestBed.configureTestingModule({
-      imports: [
-        RouterTestingModule.withRoutes([{ path: 'authority', component: AuthorityComponent }]),
-        HttpClientTestingModule,
-        AuthorityComponent,
-      ],
+      imports: [HttpClientTestingModule, AuthorityComponent],
       providers: [
         {
           provide: ActivatedRoute,
