@@ -3,7 +3,6 @@ import { HttpResponse } from '@angular/common/http';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { FormBuilder } from '@angular/forms';
 import { ActivatedRoute } from '@angular/router';
-import { RouterTestingModule } from '@angular/router/testing';
 import { of, Subject, from } from 'rxjs';
 
 import { AuthorityService } from '../service/authority.service';
@@ -21,7 +20,7 @@ describe('Authority Management Update Component', () => {
 
   beforeEach(() => {
     TestBed.configureTestingModule({
-      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([]), AuthorityUpdateComponent],
+      imports: [HttpClientTestingModule, AuthorityUpdateComponent],
       providers: [
         FormBuilder,
         {
