@@ -2,7 +2,6 @@ jest.mock('app/core/auth/account.service');
 
 import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { Router } from '@angular/router';
-import { RouterTestingModule } from '@angular/router/testing';
 import { of, Subject } from 'rxjs';
 
 import { AccountService } from 'app/core/auth/account.service';
@@ -28,7 +27,7 @@ describe('Home Component', () => {
 
   beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
-      imports: [HomeComponent, RouterTestingModule.withRoutes([])],
+      imports: [HomeComponent],
       providers: [AccountService],
     })
       .overrideTemplate(HomeComponent, '')
